reorganized files for standard flask organization
