+++ app/db.py
@@ -1,0 +1,8 @@
+import sqlite3
+
+DATABASE = 'quotes.db'
+
+def get_db_connection():
+    conn = sqlite3.connect(DATABASE)
+    conn.row_factory = sqlite3.Row
+    return conn
