+++ app/templates/index.html
@@ -2,8 +2,8 @@
 <html>
 <head>
   <title>Daily Quote</title>
-  <script defer src="script.js"></script>
 </head>
+<script src="{{ url_for('static', filename='js/script.js') }}"></script>
 <body>
   <h1>Quote of the Moment</h1>
   <p id="quote">Click the button to load a quote!</p>
